[pg] extract ArrayTypes registry class into a separate file
