+++ lib/rom/sql/extensions/postgres/types/array.rb
@@ -2,6 +2,8 @@ require 'sequel/core'
 
 Sequel.extension(*%i(pg_array pg_array_ops))
 
+require 'rom/sql/extensions/postgres/types/array_types'
+
 module ROM
   module SQL
     module Postgres
@@ -11,71 +13,8 @@ module ROM
         ArrayRead = Array.constructor { |v| v.respond_to?(:to_ary) ? v.to_ary : v }
 
         # @api private
-        class ArrayTypes
-          attr_reader :elements
-
-          attr_reader :constructor
-
-          attr_reader :base_write_type
-
-          attr_reader :base_read_type
-
-          def initialize
-            @elements = {}
-            @base_write_type = Postgres::Types::Array
-            @base_read_type = ArrayRead
-            @constructor = proc { |db_type, member|
-              -> arr {
-                if member
-                  Sequel.pg_array(arr.map { |v| member[v] }, db_type)
-                else
-                  Sequel.pg_array(arr, db_type)
-                end
-              }
-            }
-          end
-
-          def [](db_type, member_type = nil)
-            elements.fetch(db_type) do
-              name = "#{db_type}[]"
-
-              write_type = build_write_type(db_type, member_type)
-              read_type = build_read_type(db_type, member_type)
-
-              array_type = Types.Type(name, write_type).meta(type: db_type, read: read_type)
-
-              register_extension(array_type)
-
-              elements[db_type] = array_type
-            end
-          end
-
-          private
-
-          def build_write_type(db_type, member_type)
-            if member_type
-              base_write_type.constructor(constructor[db_type, member_type])
-            else
-              base_write_type.constructor(constructor[db_type])
-            end
-          end
-
-          def build_read_type(db_type, member_type)
-            if member_type && member_type.meta[:read]
-              base_read_type.of(member_type.meta[:read])
-            else
-              base_read_type
-            end
-          end
-
-          def register_extension(type)
-            TypeExtensions.register(type) { include ArrayMethods }
-          end
-        end
-
-        # @api private
         def self.array_types
-          @array_types ||= ArrayTypes.new
+          @array_types ||= ArrayTypes.new(Postgres::Types::Array, Postgres::Types::ArrayRead)
         end
 
         # @api private
